Mark ground tilemap as kinematic
Note that I'm not sure why the course decides to mark it as
Kinematic and not as Static, but I do know that high collision
velocities can result in issues with Discrete kinematic collision
detection; if this happens, it's going to be fixed by setting the
collision detection to Continuous. Static bodies don't have this
option, so maybe that's a reason.