Fix player interaction collider, add doors
Before, the interaction collider of the player was set to be
a trigger. Because of this, it couldn't "collide" with doors
in order to trigger the door opening. However, setting it to
a non-trigger resulted in collision issues with other elements
on the map, such as hazards.

This commit essentially reduces the interaction collider of the
player and makes it a non-trigger to solve both issues.

Note that the better solution is probably to move it to another
game layer and suppress collisions for that layer entirely.